refactor: extract media-processing, thumbnail, tagging services to packages/application
