+++ apps/server/src/tests/integration/media/copy-media-integration.test.ts
@@ -131,17 +131,26 @@ describe("MediaService - Copy Media Integration", () => {
 		);
 
 		services.registerMediaProcessingService(
-			new MediaProcessingServiceImpl(
-				services.getSourceRepository(),
-				services.getMediaRepository(),
-				services.getTagRepository(),
-				services.getAuthorRepository(),
-				services.getCharacterService(),
-				services.getIpRepository(),
-				services.getProjectRepository(),
-				services.getJobRepository() as any,
-				mockConfigService as any,
-			),
+			new MediaProcessingServiceImpl({
+				sourceRepo: services.getSourceRepository(),
+				mediaRepo: services.getMediaRepository(),
+				tagRepo: services.getTagRepository(),
+				authorRepo: services.getAuthorRepository(),
+				characterRepo: services.getCharacterRepository(),
+				ipRepo: services.getIpRepository(),
+				projectRepo: services.getProjectRepository(),
+				jobRepo: services.getJobRepository() as any,
+				imageProcessor: {} as any,
+				mediaStorage: {} as any,
+				enableAutoTagging: false,
+				supportedExtensions: {
+					image: [".jpg", ".jpeg", ".png", ".webp"],
+					video: [".mp4", ".webm", ".mov"],
+					audio: [".mp3", ".wav"],
+				},
+				generateThumbnail: vi.fn() as any,
+				sseSendEvent: vi.fn() as any,
+			}),
 		);
 
 		// Clean DB
