+++ packages/application/src/ports/thumbnail-service.ts
@@ -1,0 +1,15 @@
+export interface IThumbnailService {
+  getMediaThumbnailUrl(
+    mediaSourceId: string,
+    mediaId: string,
+    size?: number,
+  ): string;
+
+  startThumbnailGeneration(
+    mediaSourceId: string,
+  ): Promise<{ success: boolean; count: number }>;
+
+  clearThumbnailCache(
+    mediaSourceId: string,
+  ): Promise<{ success: boolean }>;
+}
